링크 오류 해결, Matrix3x3::Identity 가 초기화 되지 않는 문제 발생

+++ WindowsProject1/Vector2.h
@@ -39,6 +39,6 @@ public:
 
 	float Magnitude() const
 	{
-		return std::sqrtf(x * x + y * y);
+		return std::sqrt(x * x + y * y);
 	}
 };
